requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2.8
django-allauth==0.59.0
django-auto-logout==0.5.1
django-cleanup==8.0.0
django-jalali-date==1.0.2
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
httplib2==0.22.0
idna==3.4
jdatetime==4.1.1
loguru==0.7.2
mysqlclient==2.2.0
oauthlib==3.2.2
Pillow==10.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-decouple==3.8
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
win32-setctime==1.1.0
